refactor(workflow): move the step infos reset util out of the iterator
