+++ packages/twenty-server/src/modules/workflow/workflow-executor/utils/build-step-infos-reset.util.ts
@@ -1,27 +1,14 @@
 import { isDefined } from 'twenty-shared/utils';
 import { StepStatus, type WorkflowRunStepInfo } from 'twenty-shared/workflow';
 
-import { getAllStepIdsInLoop } from 'src/modules/workflow/workflow-executor/workflow-actions/iterator/utils/get-all-step-ids-in-loop.util';
-import { type WorkflowAction } from 'src/modules/workflow/workflow-executor/workflow-actions/types/workflow-action.type';
-
-export const buildLoopStepInfosReset = ({
-  iteratorStepId,
-  initialLoopStepIds,
-  steps,
+export const buildStepInfosReset = ({
+  stepIds,
   stepInfos,
 }: {
-  iteratorStepId: string;
-  initialLoopStepIds: string[];
-  steps: WorkflowAction[];
+  stepIds: string[];
   stepInfos: Record<string, WorkflowRunStepInfo>;
 }): Record<string, WorkflowRunStepInfo> => {
-  const stepIdsToReset = getAllStepIdsInLoop({
-    iteratorStepId,
-    initialLoopStepIds,
-    steps,
-  });
-
-  return stepIdsToReset.reduce(
+  return stepIds.reduce(
     (acc, stepId) => {
       const stepInfo = stepInfos[stepId];
